docs(storybook): improve argTypes and update component docs
StorybookのargTypes定義を更新し、コントロールが無効化されている複雑なpropsに説明を追加しました。

- control: { type: null } を control: false に変更
- descriptionを追加して、UI上での編集が非推奨である理由を明記
- ダイアログコンポーネントのドキュメントコメントを実装に合わせて修正
